// services/portfolio-service/src/models/clientPortal/ClientPortal.ts
export enum DashboardWidgetType {
  PORTFOLIO_SUMMARY = 'PORTFOLIO_SUMMARY',
  PERFORMANCE_CHART = 'PERFORMANCE_CHART',
  ASSET_ALLOCATION = 'ASSET_ALLOCATION',
  RECENT_TRANSACTIONS = 'RECENT_TRANSACTIONS',
  MARKET_NEWS = 'MARKET_NEWS',
  ACCOUNT_BALANCE = 'ACCOUNT_BALANCE',
  WATCHLIST = 'WATCHLIST',
  ALERTS = 'ALERTS',
  DOCUMENTS = 'DOCUMENTS',
  GOALS_PROGRESS = 'GOALS_PROGRESS'
}

export enum MessageType {
  GENERAL = 'GENERAL',
  ALERT = 'ALERT',
  STATEMENT = 'STATEMENT',
  TRADE_CONFIRMATION = 'TRADE_CONFIRMATION',
  DOCUMENT_NOTIFICATION = 'DOCUMENT_NOTIFICATION',
  SYSTEM_MAINTENANCE = 'SYSTEM_MAINTENANCE'
}

export enum MessageStatus {
  UNREAD = 'UNREAD',
  READ = 'READ',
  ARCHIVED = 'ARCHIVED',
  DELETED = 'DELETED'
}
// ...
export interface DashboardDataRequest {
  clientId: string;
  widgetTypes: DashboardWidgetType[];
  dateRange?: {
    startDate: Date;
    endDate: Date;
  };
  portfolioIds?: string[];
}